Create the ref between the contacts and user, and resolve the typo bugs

+++ controllers/contactController.js
@@ -3,7 +3,7 @@ const Contact = require("../models/contactModel")
 
 
 const getContacts = asyncHandler(async (req,res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts) 
 })
 
@@ -15,6 +15,7 @@ const createContact = asyncHandler(async (req,res)=>{
     }
 
     const contact = await Contact.create({
+        user_id: req.user.id,
         name,
         email,
         phone
@@ -39,6 +40,11 @@ const updateContact = asyncHandler(async(req,res)=>{
         throw new Error("Contact not found!")
     }
 
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contacts")
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -52,6 +58,11 @@ const deleteContact = asyncHandler(async(req,res)=>{
     if (!contact){
         res.status(404)
         throw new Error("Contact not found!")
+    }
+    
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contacts")
     }
 
     await Contact.findByIdAndDelete(req.params.id)
